requirements.txt
streamlit==1.31.0
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
numpy<2.0.0
